EmpOutcome_03 Rule and Unit Tests, updated DD07 to use int instead of long

+++ src/ESFA.DC.ILR.ValidationService.Rules/Derived/DD07.cs
@@ -7,9 +7,9 @@ namespace ESFA.DC.ILR.ValidationService.Rules.Derived
 {
     public class DD07 : IDD07
     {
-        private readonly IEnumerable<long?> _allowedProgTypes = new HashSet<long?>() { 2, 3, 20, 21, 22, 23, 25 };
+        private readonly IEnumerable<int?> _allowedProgTypes = new HashSet<int?>() { 2, 3, 20, 21, 22, 23, 25 };
 
-        public string Derive(long? progType)
+        public string Derive(int? progType)
         {
             return _allowedProgTypes.Contains(progType) ? ValidationConstants.Y : ValidationConstants.N;
         }
